chore: refactor Playwright configuration and add new E2E tests for connection settings
